Refactor and re-organise (#97)
* Refactor and re-organise

* Refactor page structure

* Refactor careers

* Remove unused document

* Kill dead code

* Remove unused import
