+++ components/Button.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 // import { transparentize } from 'polished'
-import { Defaults, Color, Font, Media } from 'const/styles/variables'
+import { Defaults, Color, Font, Media } from 'styles/variables'
 
 type ButtonProps = {
   wrapText?: boolean
@@ -15,21 +15,21 @@ type ButtonProps = {
   minHeight?: number
 }
 
-const Wrapper = styled.a<Omit<ButtonProps, "href" | "label" | "target" | "rel">>`
+const Wrapper = styled.a<Omit<ButtonProps, 'href' | 'label' | 'target' | 'rel'>>`
   display: flex;
   background: ${Color.darkBlue};
   flex-flow: row;
   border: 0.1rem solid transparent;
   color: ${Color.white};
-  padding: ${({ paddingLR }) => paddingLR ? `0 ${paddingLR}rem` : '0 2.4rem'};
+  padding: ${({ paddingLR }) => (paddingLR ? `0 ${paddingLR}rem` : '0 2.4rem')};
   box-sizing: border-box;
-  border-radius: ${({ borderRadius }) => borderRadius ? borderRadius : Defaults.borderRadius};
-  min-height: ${({ minHeight }) => minHeight ? `${minHeight}rem` : '5.6rem'};
+  border-radius: ${({ borderRadius }) => (borderRadius ? borderRadius : Defaults.borderRadius)};
+  min-height: ${({ minHeight }) => (minHeight ? `${minHeight}rem` : '5.6rem')};
   align-items: center;
-  font-size: ${({ fontSize }) => fontSize ? `${fontSize}rem` : '2.2rem'};
+  font-size: ${({ fontSize }) => (fontSize ? `${fontSize}rem` : '2.2rem')};
   justify-content: center;
   transition: color 0.2s ease-in-out, background 0.2s ease-in-out;
-  white-space: ${({ wrapText }) => wrapText ? 'initial' : 'nowrap'};
+  white-space: ${({ wrapText }) => (wrapText ? 'initial' : 'nowrap')};
   font-weight: ${Font.weightMedium};
   text-decoration: none;
 
@@ -38,54 +38,66 @@ const Wrapper = styled.a<Omit<ButtonProps, "href" | "label" | "target" | "rel">>
     min-height: 4.8rem;
   }
 
-
-  ${({ variant }) => variant === 'outline' && `
+  ${({ variant }) =>
+    variant === 'outline' &&
+    `
     background: transparent;
     border: 0.1rem solid ${Color.darkBlue};
     color: ${Color.darkBlue};
   `}
 
-  ${({ variant }) => variant === 'small' && `
+  ${({ variant }) =>
+    variant === 'small' &&
+    `
     min-height: 3.6rem;
-    border-radius: ${({ borderRadius }) => borderRadius ? borderRadius : '1.2rem'};
+    border-radius: ${({ borderRadius }) => (borderRadius ? borderRadius : '1.2rem')};
   `}
 
-  ${({ variant }) => variant === 'text' && `
+  ${({ variant }) =>
+    variant === 'text' &&
+    `
     background: transparent;
     color: ${Color.darkBlue};
   `}
 
-${({ variant }) => variant === 'textLight' && `
+${({ variant }) =>
+    variant === 'textLight' &&
+    `
     background: transparent;
     color: ${Color.lightBlue};
   `}
 
-  ${({ variant }) => variant === 'light' && `
+  ${({ variant }) =>
+    variant === 'light' &&
+    `
     background: ${Color.lightBlue};
     color: ${Color.darkBlue};
   `}
 
-  ${({ variant }) => (variant === 'outlineLight') && `
+  ${({ variant }) =>
+    variant === 'outlineLight' &&
+    `
     border: 0.1rem solid ${Color.lightBlue};
   `}
 
   &:hover {
-    background: ${({ variant }) => variant === 'outline' ? Color.darkBlue : Color.text1};
+    background: ${({ variant }) => (variant === 'outline' ? Color.darkBlue : Color.text1)};
     color: ${Color.lightBlue};
   }
 `
 
 // General purpose multiple button wrapper
-export const ButtonWrapper = styled.div<{center?: boolean}>`
+export const ButtonWrapper = styled.div<{ center?: boolean }>`
   display: flex;
   gap: 1.6rem;
   width: 100%;
 
-  ${({ center }) => center && `
+  ${({ center }) =>
+    center &&
+    `
     justify-content: center;
     align-items: center;
   `}
-
 
   ${Media.mediumDown} {
     flex-flow: column wrap;
@@ -103,14 +115,19 @@ export default function Button({
   fontSize,
   paddingLR,
   variant,
-  href = "#",
+  href = '#',
   label,
   target,
   rel,
-  minHeight
+  minHeight,
 }: ButtonProps) {
   return (
-    <Wrapper {...{ wrapText, borderRadius, fontSize, paddingLR, variant, minHeight }} href={href} target={target} rel={rel}>
+    <Wrapper
+      {...{ wrapText, borderRadius, fontSize, paddingLR, variant, minHeight }}
+      href={href}
+      target={target}
+      rel={rel}
+    >
       {label}
     </Wrapper>
   )
