+++ components/ChartSection/index.tsx
@@ -3,7 +3,7 @@ import { Chart, TimePeriod } from '../Chart'
 import { ParentSize } from '@visx/responsive'
 import { useMemo } from 'react'
 import { useQuery } from '@apollo/client'
-import { tokenPriceQuery, HistoryDuration, Chain } from 'services/graphql/queries'
+import { tokenPriceQuery, HistoryDuration, Chain } from 'services/uniswap-price/queries'
 import { usePriceHistory } from 'lib/hooks/usePriceHistory'
 import { fixChart } from 'util/fixChart'
 import { Platforms } from 'types'
