+++ components/Layout/Header.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import Link from 'next/link'
 import { transparentize, lighten } from 'polished'
 import Button from 'components/Button'
-import { Defaults, Color, Font, Media } from 'const/styles/variables'
+import { Defaults, Color, Font, Media } from 'styles/variables'
 import { InView } from 'react-intersection-observer'
 import useMediaQuery from 'lib/hooks/useMediaQuery'
 import { useRouter } from 'next/router'
