+++ styles/variables.ts
